tts/xtts/requirements.txt
# XTTS Engine Dependencies
#
# NOTE: Uses coqui-tts (maintained Idiap fork) instead of abandoned TTS package
# See README.md for detailed versioning decisions

# Core TTS - Maintained fork of Coqui TTS (original TTS package abandoned Dec 2023)
coqui-tts>=0.27.0,<0.28.0

# PyTorch - Pinned to 2.5.x for CUDA 11.8/12.1/12.4 compatibility
# coqui-tts 0.27.x requires: torch >2.1, <2.9
torch>=2.5.0,<2.6.0
torchaudio>=2.5.0,<2.6.0

# Transformers - Version range required by coqui-tts 0.27.x
# coqui-tts 0.27.x requires: transformers >4.52.1, <4.56
transformers>=4.53.0,<4.56.0

# FastAPI Server Stack
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
pydantic>=2.10.0,<3.0.0
python-multipart>=0.0.9,<1.0.0

# Utilities
loguru>=0.7.2,<0.8.0
httpx>=0.28.0,<1.0.0
PyYAML>=6.0.0,<7.0.0
huggingface_hub>=0.20.0

# CUDA support (optional, comment out for CPU-only)
# Supports: CUDA 11.8, 12.1, 12.4
--extra-index-url https://download.pytorch.org/whl/cu118
